Integrate Liquibase

+++ src/main/java/io/github/JRojowski/entity/dto/RecipeDto.java
@@ -5,5 +5,5 @@ import lombok.Getter;
 @Getter
 public class RecipeDto {
     private FoodDto foodDto;
-    private float grams;
+    private double grams;
 }
